Update heatmap.2.R
Added a mgps parameter that loads the mgp argument in par() function and, additionally, two xmgp and ymgp parameters to control specifically for the first field in mgp for x and y independently. Therefore, if not used they just take mgps[1] as value, otherwise
these parametes overwrite mgps[1]. In general, the default values for mgps work very well, but sometimes it is needed to modify it to avoid overlaps.